% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_covars.R
\name{compare_covars}
\alias{compare_covars}
\title{Covariate Comparison}
\usage{
compare_covars(
  redd_df = NULL,
  species = c("Steelhead", "Spring Chinook"),
  num_obs = c("two", "one")
)
}
\arguments{
\item{redd_df}{dataframe containing redd data including columns with covariates in selected observer error model}

\item{species}{which species is being analyzed? This function chooses the appropriate error model for this species. Choices are \code{Steelhead} or \code{Chinook}, with \code{Steelhead} being the default.}

\item{num_obs}{if species is \code{Steelhead}, which error model to use, for one or two observers? Default is \code{two}.}
}
\value{
dataframe
}
\description{
Compares covariates of model dataset with predictive dataset
}
\author{
Kevin See
}
