% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_rchs.R
\name{correlate_rchs}
\alias{correlate_rchs}
\title{Estimate Correlation Between Reaches}
\usage{
correlate_rchs(
  redd_df = NULL,
  date_nm = "SurveyDate",
  cor_redd_nm = "NewRedds",
  reach_nm = "Reach",
  use = "pairwise.complete.obs",
  ...
)
}
\arguments{
\item{redd_df}{dataframe containing redd data, including estimates of observer error}

\item{date_nm}{quoted name of column in {redd_df} listing date of each survey}

\item{cor_redd_nm}{quoted name of column in {redd_df} listing number of redds found during that survey, to be used in the correlation}

\item{reach_nm}{quoted name of column in {redd_df} listing name of reach}
}
\value{
dataframe
}
\description{
Estimates correlation between reaches, to be used when calculating standard error across multiple reaches so reaches aren't necessarily treated independently
}
\author{
Kevin See
}
