% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_err_fncs.R
\name{est_adj_prop}
\alias{est_adj_prop}
\title{Adjust Observation-Error Proportions}
\usage{
est_adj_prop(boot_df = NULL, estimate = c("x_y_ratio", "prop_x"), ...)
}
\arguments{
\item{boot_df}{bootstrap results of calling the function \code{boot_adj_prop()}.}

\item{estimate}{if \code{results == estimate}, should an estimate of the ratio of X:Y be returned (\code{x_y_ratio}), or an estimate of the proportion of X (\code{prop_x})? Default is \code{x_y_ratio}.}
}
\value{
a data.frame
}
\description{
This function simulates some data from a scenario when initial sex calls are error-prone to some degree, but different data streams can provide insight into what the sex call error rates are. It is meant to mimic data from an initial sample, like adult steelhead sampling at Priest Rapids Dam, with sex-specific error rates, and subsequent broodstock collections of some of those fish later, when sex calls are known without error.
}
\author{
Kevin See
}
