% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gauc.R
\name{fit_gauc}
\alias{fit_gauc}
\title{Gaussian Area-Under-the-Curve}
\usage{
fit_gauc(
  data,
  SL = 1,
  SL_se = 0,
  v = 1,
  v_se = 0,
  model_fam = c("quasipoisson", "negative.binomial")
)
}
\arguments{
\item{data}{dataframe containing columns named redds and day (Julian days)}

\item{v}{overall mean observer error}

\item{v_se}{standard deviation of overall observer error}

\item{model_fam}{model family. Must be either \code{quasipoisson} or \code{negative.binomial}.}
}
\value{
list
}
\description{
Fit a Gaussian area-under-the-curve model to redd counts, incorporating estimates of observer error if possible.
}
\author{
Kevin See
}
