% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_obs_sthd}
\alias{one_obs_sthd}
\title{One Observer Steelhead Redd Observer Error}
\format{
A data frame with 100 rows and 74 variables:
\describe{
\item{year}{Spawn year}
\item{stream}{What stream is the survey in}
\item{reach}{what reach was surveyed}
\item{surveyor}{code for which surveyor conducted survey}
\item{date}{date of survey}
\item{visible_redds}{how many redds were visible to be counted on that date, based on census surveys}
\item{total_features}{how many redds did the naive surveyors count on that survey}
...
}
}
\source{
Michael Hughes and Andrew Murdoch at WDFW
}
\usage{
one_obs_sthd
}
\description{
A dataset containing the data to fit the one-observer steelhead redd observer error model
}
\keyword{datasets}
