% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_neterr.R
\name{predict_neterr}
\alias{predict_neterr}
\title{Predict Net Error}
\usage{
predict_neterr(
  redd_df,
  species = c("Steelhead", "Spring Chinook"),
  num_obs = c("two", "one")
)
}
\arguments{
\item{redd_df}{dataframe containing redd data including columns with covariates in selected observer error model}

\item{species}{which species is being analyzed? This function chooses the appropriate error model for this species. Choices are \code{Steelhead} or \code{Chinook}, with \code{Steelhead} being the default.}

\item{num_obs}{if species is \code{Steelhead}, which error model to use, for one or two observers? Default is \code{two}.}
}
\value{
tibble
}
\description{
Predict the net error of a steelhead or Spring Chinook redd survey, based on published model
}
\author{
Kevin See
}
