% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_wen_sthd_data.R
\name{prep_wen_sthd_data}
\alias{prep_wen_sthd_data}
\title{Prepare All Data from Wenatchee for a Report}
\usage{
prep_wen_sthd_data(
  redd_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/inputs/Redd Data",
  redd_file_name = "Wenatchee_Redd_Surveys.xlsx",
  experience_path = redd_file_path,
  experience_file_name = redd_file_name,
  dabom_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/estimates",
  dabom_file_name = "UC_STHD_Model_Output.xlsx",
  brood_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/inputs/Bio Data/Sex and Origin PRD-Brood Comparison Data",
  brood_file_name = "STHD UC Brood Collections_2011 to current.xlsx",
  removal_file_path =
    "T:/DFW-Team FP Upper Columbia Escapement - General/UC_Sthd/inputs/Fish Removals",
  removal_file_name = "Master_STHD_Removals_2.18.23.MH.xlsx",
  n_observers = "two",
  query_year = lubridate::year(lubridate::today()) - 1,
  phos_data = c("tags", "escapement"),
  save_rda = F,
  save_by_year = T,
  save_file_path = here::here("analysis/data/derived_data"),
  save_file_name = NULL
)
}
\arguments{
\item{redd_file_path}{file path to redd data file}

\item{redd_file_name}{name of Excel file containing redd data in a very particular format, including sheets labeled "Reach Length", "Thalweg CV", "Discharge Gages" and "Redd Surveys".}

\item{experience_path}{file path to experience file}

\item{experience_file_name}{name of Excel file containing experience data in a very particular format, with a sheet labeled "Experience"}

\item{dabom_file_path}{file path to summary results from DABOM}

\item{dabom_file_name}{name of Excel file containing DABOM estimates in a very particular format}

\item{brood_file_path}{file path to broodstock collection file}

\item{brood_file_name}{name of Excel file containing broodstock collection data in a very particular format}

\item{removal_file_path}{file path to removal data file}

\item{removal_file_name}{name of Excel file containing removal data in a very particular format}

\item{n_observers}{how many observers / boats were used on each survey?}

\item{query_year}{which year or years should be included in this query?}

\item{phos_data}{should the data used to estimate pHOS come from PIT \code{tags} or \code{escapement} estimates? Default is \code{tags}.}

\item{save_rda}{should the data that's returned by saved as an .RData object (\code{TRUE})? Default value of \code{FALSE} loads all the returned objects into the Global Environment.}

\item{save_file_path}{if \code{save_rda} is \code{TRUE}, where should the .RData object be saved?}

\item{save_file_name}{if \code{save_rda} is \code{TRUE}, what should the file name be? Should end in ".rda".}
}
\value{
either saves an .Rdata object with output for a given year, or loads those results into global environment.
}
\description{
This function pulls together redd survey data and estimates net error for each survey, PIT tags detected in the Wenatchee, error rates of sex calls at Priest, a corrected fish/redd estimate based on sex ratios adjusted for the error rate of sex calls at Priest, estimates of escapement to various tributaries in the Wenatchee as well as run escapement to the Wenatchee as a whole and the number of known removals (e.g. broodstock collection, harvest). These are all either saved as an .RData object, or loaded into the global environment.
}
\author{
Kevin See
}
