% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mod_summ.R
\name{print_mod_summ}
\alias{print_mod_summ}
\title{Print Model Summary}
\usage{
print_mod_summ(
  species = c("Steelhead", "Spring Chinook"),
  num_obs = c("two", "one")
)
}
\arguments{
\item{species}{which species is being analyzed? This function chooses the appropriate error model for this species. Choices are \code{Steelhead} or \code{Chinook}, with \code{Steelhead} being the default.}

\item{num_obs}{if species is \code{Steelhead}, which error model to use, for one or two observers? Default is \code{two}.}
}
\value{
dataframe
}
\description{
Shows estimated coefficients for selected observer error model
}
\author{
Kevin See
}
