% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_err_fncs.R
\name{sim_sex_err}
\alias{sim_sex_err}
\title{Simulate sexing error data}
\usage{
sim_sex_err(
  n_samp = 1,
  p_true = NULL,
  x_err = NULL,
  y_err = NULL,
  b_x = NULL,
  b_y = NULL,
  n_sim = 1,
  seed = NA
)
}
\arguments{
\item{n_samp}{Total number of individuals in initial sample}

\item{p_true}{True proportion of group X}

\item{x_err}{Error rate of identifying group X. Of all the actual individuals in group X, this proportion will be identified as group Y.}

\item{y_err}{Error rate of identifying group Y. Of all the actual individuals in group Y, this proportion will be identified as group X.}

\item{b_x}{Number of group X individuals in second sample (e.g., broodstock collection)}

\item{b_y}{Number of group Y individuals in second sample (e.g., broodstock collection)}

\item{n_sim}{How many simulated data sets should be created? Default value is \code{1}}

\item{seed}{if set to a numeric value, it invokes the \code{set.seed()} function to make the data reproducible}
}
\value{
a data.frame
}
\description{
This function simulates some data from a scenario when initial sex calls are error-prone to some degree, but different data streams can provide insight into what the sex call error rates are. It is meant to mimic data from an initial sample, like adult steelhead sampling at Priest Rapids Dam, with sex-specific error rates, and subsequent broodstock collections of some of those fish later, when sex calls are known without error.
}
\author{
Kevin See
}
