% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sroem-package.R
\docType{package}
\name{sroem-package}
\alias{sroem}
\alias{sroem-package}
\title{sroem: Salmonid Redd Observer Error Models}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinSee/sroem}
  \item Report bugs at \url{https://github.com/KevinSee/sroem/issues}
}

}
\author{
\strong{Maintainer}: Kevin See \email{Kevin.See@dfw.wa.gov} (\href{https://orcid.org/0000-0002-9762-6442}{ORCID})

}
\keyword{internal}
