% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_redds.R
\name{summarize_redds}
\alias{summarize_redds}
\title{Summarize Redds}
\usage{
summarize_redds(
  redd_df = NULL,
  species = c("Steelhead", "Spring Chinook"),
  group_vars = c(river, reach, index, survey_type),
  summ_vars = c("river", "location"),
  new_redd_nm = "new_redds",
  vis_redd_nm = "visible_redds",
  net_err_nm = "net_error",
  net_se_nm = "net_error_se",
  min_non0_wks = 3,
  min_redds = 2,
  gauc = NULL,
  add_zeros = F,
  use_cor = F,
  date_nm = "survey_date",
  cor_redd_nm = "new_redds",
  reach_nm = "reach",
  use = "pairwise.complete.obs",
  method = c("pearson", "kendall", "spearman"),
  ...
)
}
\arguments{
\item{redd_df}{dataframe containing redd data, including estimates of observer error}

\item{species}{which species is being analyzed? This function chooses the appropriate error model for this species. Choices are \code{Steelhead} or \code{Chinook}, with \code{Steelhead} being the default.}

\item{group_vars}{vector of column names from {redd_df} to group results by}

\item{summ_vars}{vector of column names from {redd_df} to group summaries of redd results by}

\item{new_redd_nm}{quoted name of column in {redd_df} listing number of new redds found during that survey}

\item{vis_redd_nm}{quoted name of column in {redd_df} listing number of visible redds present during that survey}

\item{net_err_nm}{quoted name of column in {redd_df} listing estimate of net error for that survey}

\item{net_se_nm}{quoted name of column in {redd_df} listing standard error of net error estimate for that survey}

\item{min_non0_wks}{minimum number of weeks with at least one new redd observed}

\item{min_redds}{minimum number of total redds observed}

\item{gauc}{\code{TRUE/FALSE} should a Gaussian area-under-the-curve model be applied? If \code{FALSE}, the total number of new redds is divided by the mean estimate of observer error. By default \code{gauc} is \code{TRUE} for steelhead, and \code{FALSE} for Spring Chinook.}

\item{add_zeros}{should leading and trailing zero counts be added if the first and last counts aren't already zero? Default is {FALSE}.}

\item{use_cor}{should correlations between reaches be used in calculating standard errors? Default is {FALSE}}

\item{date_nm}{quoted name of column in {redd_df} listing date of each survey}

\item{cor_redd_nm}{quoted name of column in {redd_df} listing number of redds found during that survey, to be used in the correlation}

\item{reach_nm}{quoted name of column in {redd_df} listing name of reach}
}
\value{
dataframe
}
\description{
Estimates true number of redds across a series of reaches, using a Gaussian area-under-the-curve model where possible and incorporating estimates of observer error if possible. These estimates are then summarized across a user-defined spatial scale.
}
\author{
Kevin See
}
