% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_err_fncs.R
\name{est_adj_prop}
\alias{est_adj_prop}
\title{Adjust Observation-Error Proportions}
\usage{
est_adj_prop(
  b_x,
  b_y,
  false_x,
  false_y,
  samp_x,
  samp_y,
  n_boot = 1000,
  seed = NA,
  results = c("estimate", "bootstrap")
)
}
\arguments{
\item{b_x}{Number of group X individuals in second sample (e.g., broodstock collection)}

\item{b_y}{Number of group Y individuals in second sample (e.g., broodstock collection)}

\item{false_x}{how many individuals from group \code{b_x} were mis-identified as Y's in initial sample?}

\item{false_y}{how many individuals from group \code{b_y} were mis-identified as X's in initial sample?}

\item{samp_x}{number of individuals in initial sample identified as X.}

\item{samp_y}{number of individuals in initial sample identified as Y.}

\item{n_boot}{how many bootstrap samples should be used? Default is \code{1000}.}

\item{seed}{if set to a numeric value, it invokes the \code{set.seed()} function to make the data reproducible}

\item{results}{should this function return a dataframe of estimates based on bootstrap samples (\code{estimate}), or a dataframe of those bootstrap samples (\code{bootstrap})? Default is \code{estimate}.}
}
\value{
a data.frame
}
\description{
This function simulates some data from a scenario when initial sex calls are error-prone to some degree, but different data streams can provide insight into what the sex call error rates are. It is meant to mimic data from an initial sample, like adult steelhead sampling at Priest Rapids Dam, with sex-specific error rates, and subsequent broodstock collections of some of those fish later, when sex calls are known without error.
}
\author{
Kevin See
}
