% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dabom_results.R
\name{query_dabom_results}
\alias{query_dabom_results}
\title{Query DABOM results}
\usage{
query_dabom_results(
  dabom_file_path =
    "O:Documents/Git/MyProjects/DabomPriestRapidsSthd/analysis/data/derived_data/estimates",
  dabom_dam_nm = c("RockIsland", "PriestRapids"),
  dabom_file_name = "UC_Sthd_DABOM_",
  query_year = lubridate::year(lubridate::today()) - 1,
  result_type = c("escape_summ", "tag_summ", "escape_post", "dam_summ", "total_escape",
    "detect_summ", "parent_child")
)
}
\arguments{
\item{dabom_file_path}{file path to DABOM results. Should contain at least two folders names \code{RockIsland} and \code{PriestRapids}.}

\item{dabom_dam_nm}{name of dam where dam counts were used to generate escapement estimates. Choices are \code{RockIsland} (default) or \code{PriestRapids}}

\item{dabom_file_name}{partial name of .Rdata object saved after running DABOM model for Upper Columbia.}

\item{query_year}{which spawn year should results be queried fof?}

\item{result_type}{what type of results should be queried for? Options include summary of escapement estimates (\code{escape_summ}), tag summary (\code{tag_summ}), posterior draws of escapement (\code{escape_post}), summary of dam counts and adjusted counts (\code{dam_summ}), total escapement by origin used in the results (\code{total_escape}), summary of detection probabilities (\code{detect_summ}) or parent-child table (\code{parent_child}). Default is \code{escape_summ}.}
}
\value{
tibble
}
\description{
Loads .Rdata objects that have been saved for a specific year's DABOM model run, and returns specified type of results.
}
\author{
Kevin See
}
